Punjab_Crop_Advisory/requirements_consolidated.txt
# Punjab Crop Advisory System - Consolidated Requirements
# This file consolidates all requirements from multiple requirement files
# Install with: pip install -r requirements_consolidated.txt

# ============================================
# Web Frameworks and API
# ============================================
# Flask Web Framework
Flask>=2.3.0
Flask-CORS>=3.0.0
Werkzeug>=2.3.0
Jinja2>=3.1.0
MarkupSafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
blinker>=1.6.0

# FastAPI Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Production Server
gunicorn>=20.0.0

# ============================================
# Data Science and Machine Learning
# ============================================
# Core Data Science
numpy>=1.21.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Machine Learning Models
xgboost>=1.7.0
lightgbm>=4.0.0

# ============================================
# Geospatial and Earth Engine
# ============================================
earthengine-api>=0.1.384
geemap>=0.29.0
folium==0.14.0

# ============================================
# HTTP and API Utilities
# ============================================
requests==2.31.0

# ============================================
# Development and Utilities
# ============================================
python-dotenv>=0.19.0
python-dateutil==2.8.2
pytz==2023.3
pickle-mixin==1.0.2

# ============================================
# Optional/Development Tools
# ============================================
pyngrok==7.0.0
